refactor: use big decimals instead of double for price

+++ src/main/kotlin/ecommerce/mapper/CartItemMapper.kt
@@ -10,6 +10,6 @@ fun CartItem.toDto(): CartItemResponse =
         product.id,
         productName = product.name,
         quantity = quantity,
-        productPrice = product.price,
+        productPrice = product.price.toDouble(),
         productImageUrl = product.imageUrl,
     )
